feat: Re-introduce the "create viz" CTA for homepage intro
- Browsing autocomplete is extracted and kept, it could serve later
